pyproject.toml
[project]
name = "chatboq-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "alembic>=1.16.2",
    "asyncio-redis>=0.16.0",
    "authlib>=1.6.0",
    "bcrypt>=4.3.0",
    "celery[redis]>=5.5.3",
    "fastapi[standard]>=0.115.14",
    "flower>=2.0.1",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "websockets>=15.0.1",
    "confluent-kafka>=2.3.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "httpx>=0.28.1",
    "requests>=2.32.4",
    "psycopg2-binary>=2.9.10",
    "pyotp>=2.9.0",
    "python-socketio>=5.13.0",
    "greenlet>=3.2.3",
    "python-multipart>=0.0.20",
    "cloudinary>=1.44.1",
    "redis>=5.0.0",
    "standard-distutils>=3.11.9",
    "asgiref>=3.9.1",
    "apscheduler>=3.11.0",
    "jinja2>=3.1.6",
    "sendgrid>=6.12.4",
    "arq>=0.26.3",
    "debugpy>=1.8.16",
    "pyjwt>=2.10.1",
    "dramatiq[asyncio]>=1.18.0",
    "pylint>=3.3.8",
    "cachetools>=6.1.0",
    "broadcaster>=0.3.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
